Fix derpy filesystem stuff
